일단 잡히는 W 에러들 전부 처리하였습니다.

+++ srcs/channel/Channel.cpp
@@ -3,7 +3,6 @@
 #include "Channel.hpp"
 #include "Client.hpp"
 #include "MyLibft.hpp"
-#include "ircError.hpp"
 #include <iostream>
 #include <map>
 #include <set>
@@ -133,7 +132,7 @@ bool Channel::addMember( Client &client, const std::string &passwd )
 		Msg().errBadChannelKey(client.getNickName(), m_channelName).sendTo(client, m_server);
 		return (false);
 	}
-	if (m_maxMembers <= m_members.size())
+	if (m_maxMembers <= static_cast<int>(m_members.size()))
 	{
 		Msg().errChannelIsFull(client.getNickName(), m_channelName).sendTo(client, m_server);
 		return (false);
